Refactor into web server
The reasoning is that in the future it might be desirable to interact
with the Mitigation Manager in some way, perhaps to change runtime
configuration of some sort.  By turning the manager into a web server,
we can get ahead of that use case while at the same time adding
very little additional complexity (most of the new files added have to
do with separation of concerns rather than Sanic-related code).

Signed-off-by: Ekam Puri Nieto <ekam.purin@um.es>