Overhaul attack front measurement, use custom DB exceptions
- The AttackNodes table no longer has any fleeting state associated
with it, meaning that the rows within it should only change when
CRUDing attack nodes.
- The AttackGraphs table replaces its old ongoing field with an
attack_front field that directly points to a node.
- The meaning of "ongoing/attack front" has changed from "the latest
triggered node" to "the next node in line for the attack".  This
changes how a large portion of the database methods interact with the
database, as well as the update() function's logic - it now first adds
new attack graphs to the local state, and then progresses said
state (thus progressing the new attack graphs without implicit updates
being done by database retrieval methods).

Signed-off-by: Ekam Puri Nieto <ekam.purin@um.es>