Add Hy-based condition checking
The idea is to store portable Hy code within condition objects which
can be ran alongside parameters (f.e. alert data) and data
sources (f.e. ISIM).  Hy is chosen mainly because the alternatives are
either maintaining a long list of static check functions that are
essentially written ad-hoc for whatever specific tree, or implementing
a JSON DSL from scratch (which would end up either being the same as
the ad-hoc function method, or a neverending implementation that'll
always be worse than using a real language).  Hy thus stands out as an
easily serializable format that interoperates seamlessly with the
existing Python codebase.

Signed-off-by: Ekam Puri Nieto <ekam.purin@um.es>