Fix incorrect parsing of Attack objects from SQL rows
Previously, I was relying on get_full_attack_graph() to retrieve the
entire graph based on the attack front node.  However, this method
takes as an input the initial node, and so it doesn't check for any
previous nodes when creating the graph - leading in this case to an
incomplete graph.  The solution is to search the database for the
actual initial node and generate it from there.