Fix undefined retrieve_full_graph behavior with multiple attacks
Previously, the "ongoing node" was not properly set whenever there
were multiple attacks associated with the attack graph (mainly because
there was no way of indicating which attack to choose).  Now, the
method takes an optional attack identifier to specify which attack
front to choose (when the parameter is None, the initial node is
chosen).