Fix retrieve_new_graphs not properly checking attacks
Previously, the method accessed Alert fields directly to retrieve the
MITRE Technique IDs instead of using the techniques method.
Furthermore, the way to check if an attack was already tracked was
left mostly as a stub.  This commit implements this functionality by
checking the attack's context and seeing if the alert already exists
within it.